Introduce IdGenerators.
Optimizers will need a way to reuse previous IDs generated for macros. This introduces two separate ID generation mechanism, with MonotonicIdGenerator for general use cases and StableIdGenerator intended to be used for optimizers.

The methods to produce next IDs or to renumber them will not be made public until the optimizers have been implemented.

PiperOrigin-RevId: 562830563